Remove unnecessary (long) casts in GapScanner.scanClassEnd
Fix PMD `UnnecessaryCast` warnings at lines 215 and 279 by relying
on binary numeric promotion when adding `fromPos` to `long` product.

MARKDOWN=1
TAG=agy
CONV=3b3c6460-a107-4618-b94c-d6c4f59f5e8a

+++ safere/src/main/java/org/safere/GapScanner.java
@@ -243,7 +243,7 @@ final class GapScanner {
       return cur;
     }
     if (gap.maxLength() != Integer.MAX_VALUE) {
-      return (int) Math.min(maxPos, (long) fromPos + (long) gap.maxLength() * 2);
+      return (int) Math.min(maxPos, fromPos + (long) gap.maxLength() * 2);
     }
     return maxPos;
   }
@@ -303,7 +303,7 @@ final class GapScanner {
       return cur;
     }
     if (gap.maxLength() != Integer.MAX_VALUE) {
-      return (int) Math.min(maxPos, (long) fromPos + (long) gap.maxLength() * 4);
+      return (int) Math.min(maxPos, fromPos + (long) gap.maxLength() * 4);
     }
     return maxPos;
   }
